requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
black==21.7b0
click==8.0.1
coverage==5.5
Django==3.2.5
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
pluggy==0.13.1
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytz==2021.1
regex==2021.7.6
sqlparse==0.4.1
toml==0.10.2
tomli==1.0.4
